♻️ Move `context` under `#core` (#34927)
* src/context/contextprops -> #preact/contextprops

* Remove non-core dependency from context/node

* src/context -> src/core/context

* Update imports from context

* Remove special-casing of context in eslint

* Remove special-casing of context in typechecking

* Move test files to new directory structure

* No shame(.extern.js)
