+++ extensions/amp-accordion/1.0/test/test-amp-accordion.js
@@ -16,9 +16,9 @@
 import '../amp-accordion';
 import {ActionInvocation} from '#service/action-impl';
 import {ActionTrust} from '#core/constants/action-constants';
-import {CanRender} from '../../../../src/context/contextprops';
+import {CanRender} from '#preact/contextprops';
 import {htmlFor} from '#core/dom/static-template';
-import {subscribe, unsubscribe} from '../../../../src/context';
+import {subscribe, unsubscribe} from '#core/context';
 import {toggleExperiment} from '#experiments';
 import {waitFor} from '#testing/test-helper';
