+++ extensions/amp-base-carousel/1.0/carousel-props.js
@@ -15,7 +15,7 @@
  */
 
 import {CarouselContext} from './carousel-context';
-import {contextProp} from '../../../src/context';
+import {contextProp} from '#core/context';
 
 const CarouselContextProp = contextProp('base-carousel:1.0:context', {
   type: CarouselContext,
