+++ src/core/context/node.js
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
+import * as mode from '#core/mode';
 import {Values} from './values';
 import {devAssert, devAssertElement} from '#core/assert';
-import {getMode} from '../mode';
 import {pushIfNotExist, removeItem} from '#core/types/array';
 import {throttleTail} from './scheduler';
 
@@ -66,7 +66,7 @@ export class ContextNode {
     let contextNode = /** @type {!ContextNode|undefined} */ (node[NODE_PROP]);
     if (!contextNode) {
       contextNode = new ContextNode(node, null);
-      if (getMode().localDev || getMode().test) {
+      if (mode.isLocalDev() || mode.isTest()) {
         // The `Object.defineProperty({enumerable: false})` helps tests, but
         // hurts performance. So this is only done in a dev/test modes.
         Object.defineProperty(node, NODE_PROP, {
